make edit page more reusable (#1015)
* make edit page more reusable

* update snapshots

* address comments

* update snapshot
