+++ src/components/WorkOrder/EditWorkOrder.tsx
@@ -8,19 +8,22 @@ import CharacterCountLimitedTextArea from '../Form/CharacterCountLimitedTextArea
 import Spinner from '../Spinner'
 import ErrorMessage from '../Errors/ErrorMessage'
 
-import {
-  EditWorkOrderProps,
-  FormValues,
-} from '../../types/edit-workorder/types'
-
 import { WorkOrder } from '@/models/workOrder'
 
 import {
   getWorkOrder,
   postNote,
-  editDescription,
+  editWorkOrder,
 } from '@/utils/requests/workOrders'
 import { buildNoteFormData } from '../../utils/hact/jobStatusUpdate/notesForm'
+
+export type EditWorkOrderProps = {
+  workOrderReference: string
+}
+
+export type FormValues = {
+  editRepairDescription: string
+}
 
 const EditWorkOrder = ({ workOrderReference }: EditWorkOrderProps) => {
   const [workOrder, setWorkOrder] = useState<WorkOrder | null>(null)
@@ -52,12 +55,12 @@ const EditWorkOrder = ({ workOrderReference }: EditWorkOrderProps) => {
       note: `Description updated: ${data.editRepairDescription}`,
       workOrderReference: workOrder.reference,
     })
-    const editDescriptionResponse = await editDescription(
+    const editWorkOrderResponse = await editWorkOrder(
       workOrder.reference,
       data.editRepairDescription
     )
-    if (!editDescriptionResponse.success) {
-      setError(editDescriptionResponse.error)
+    if (!editWorkOrderResponse.success) {
+      setError(editWorkOrderResponse.error)
       return
     }
 
@@ -89,7 +92,7 @@ const EditWorkOrder = ({ workOrderReference }: EditWorkOrderProps) => {
           </h1>
           <form
             role="form"
-            id="edit-description-form"
+            id="edit-work-order-form"
             onSubmit={handleSubmit(onSubmit)}
           >
             <CharacterCountLimitedTextArea
