--- src/types/edit-workorder/types.ts
@@ -1,16 +1,0 @@
-export type EditWorkOrderProps = {
-  workOrderReference: string
-}
-
-export type FormValues = {
-  editRepairDescription: string
-}
-
-export type NoteData = {
-  relatedWorkOrderReference: {
-    id: string
-  }
-  comments: string
-  typeCode: string
-  otherType: string
-}
