+++ src/types/requests/types.ts
@@ -3,3 +3,12 @@ export type ApiResponseType<T> = {
   response: T | null
   error: string | null
 }
+
+export type NoteDataType = {
+  relatedWorkOrderReference: {
+    id: string
+  }
+  comments: string
+  typeCode: string
+  otherType: string
+}
